Bug caused by database id being long instead of int fixed + some Code analysis corrections

+++ app/src/main/java/ar/edu/unlp/info/hermescelascolus/models/dao/KidDao.java
@@ -102,7 +102,7 @@ public class KidDao extends GenericDao implements Dao<Kid> {
     }
 
     @Override
-    public Kid getById(int id) {
+    public Kid getById(long id) {
 
         this.open();
         try {
